% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_by_election.R
\docType{data}
\name{by_election_2021}
\alias{by_election_2021}
\title{서울, 부산 시장 재보궐선거}
\format{
A data frame with 22589 rows and 10 variables:
\describe{
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{구분}}{character COLUMN_DESCRIPTION}
\item{\code{선거인수}}{double COLUMN_DESCRIPTION}
\item{\code{투표수}}{double COLUMN_DESCRIPTION}
\item{\code{무효투표수}}{double COLUMN_DESCRIPTION}
\item{\code{기권수}}{double COLUMN_DESCRIPTION}
\item{\code{후보}}{character COLUMN_DESCRIPTION}
\item{\code{득표수}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://info.nec.go.kr/main/main_load.xhtml}
}
\usage{
by_election_2021
}
\description{
2021년 4월 7일 재보궐선거
}
\details{
2021년 서울/부산 재보궐선거
}
\keyword{datasets}
