% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_go_kr_casting.R
\docType{data}
\name{casting_votes}
\alias{casting_votes}
\title{선거별 구시군별 투표수}
\format{
A data frame with 3032 rows and 18 variables:
\describe{
\item{\code{선거코드}}{character COLUMN_DESCRIPTION}
\item{\code{선거명}}{character COLUMN_DESCRIPTION}
\item{\code{선거구분}}{character COLUMN_DESCRIPTION}
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군명}}{character COLUMN_DESCRIPTION}
\item{\code{결과순서}}{character COLUMN_DESCRIPTION}
\item{\code{선거ID}}{character COLUMN_DESCRIPTION}
\item{\code{선거종류코드}}{character COLUMN_DESCRIPTION}
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군명}}{character COLUMN_DESCRIPTION}
\item{\code{총 선거인수}}{character COLUMN_DESCRIPTION}
\item{\code{선거일투표 선거인수}}{character COLUMN_DESCRIPTION}
\item{\code{거소·사전·선상·재외 선거인수}}{character COLUMN_DESCRIPTION}
\item{\code{총 투표자수}}{character COLUMN_DESCRIPTION}
\item{\code{선거일 투표자수}}{character COLUMN_DESCRIPTION}
\item{\code{거소·사전·선상·재외 투표자수}}{character COLUMN_DESCRIPTION}
\item{\code{투표율}}{character COLUMN_DESCRIPTION}
\item{\code{순서}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
casting_votes
}
\description{
선거별 투표결과
}
\keyword{datasets}
