% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_varnames}
\alias{clean_varnames}
\title{clean_varnames}
\usage{
clean_varnames(raw_data)
}
\arguments{
\item{raw_data}{dataframe or tibble}
}
\value{
robust unicoded dataframe or tibble
}
\description{
데이터 팩키지 변수명 한글깨짐 문제 윈도우 혹은 맥 해결
}
