% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_go_kr.R
\docType{data}
\name{code_gusigun}
\alias{code_gusigun}
\title{중앙선거관리위원회 구시군코드

공공데이터포털 - 중앙선거관리위원회 구시군코드
검색어: 중앙선거관리위원회 코드정보}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{선거코드}{문자형, 선거실시일을 코드기준값으로 설정}
\item{data}{티블, 선거코드, 시도명, 구시군명 총 3 변수를 갖는 데이터프레임}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
code_gusigun
}
\description{
중앙선거관리위원회 구시군코드
}
\examples{
code_gusigun
}
\keyword{datasets}
