% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_go_kr_master_code.R
\docType{data}
\name{code_precinct}
\alias{code_precinct}
\title{중앙선거관리위원회 선거구 코드}
\format{
63개 관측점과 4개 변수를 갖는 데이터프레임
\describe{
\item{선거코드}{문자형, 선거실시일을 코드기준값으로 설정}
\item{선거명}{문자형, 선거명칭}
\item{선거구분}{문자형, 2자리 숫자로 구성된 선거구분코드}
\item{data}{티블, 선거코드, 선거명, 시도명, 구시군명 총 4 변수를 갖는 데이터프레임}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
code_precinct
}
\description{
중앙선거관리위원회 선거구코드

공공데이터포털 - 중앙선거관리위원회 선거구코드
검색어: 중앙선거관리위원회 코드정보
}
\examples{
code_precinct
}
\keyword{datasets}
