% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_superintendent.R
\docType{data}
\name{education_2010}
\alias{education_2010}
\title{제5회 시도 교육감}
\format{
A data frame with 42874 rows and 9 variables:
\describe{
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{선거인수}}{double COLUMN_DESCRIPTION}
\item{\code{투표수}}{double COLUMN_DESCRIPTION}
\item{\code{무효투표수}}{double COLUMN_DESCRIPTION}
\item{\code{기권수}}{double COLUMN_DESCRIPTION}
\item{\code{후보}}{character COLUMN_DESCRIPTION}
\item{\code{득표수}}{double COLUMN_DESCRIPTION}
}
}
\usage{
education_2010
}
\description{
시도 교육감
}
\details{
DETAILS
}
\keyword{datasets}
