% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_superintendent.R
\docType{data}
\name{education_2014}
\alias{education_2014}
\title{제6회 교육감선거}
\format{
A data frame with 58802 rows and 10 variables:
\describe{
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{구분}}{character COLUMN_DESCRIPTION}
\item{\code{선거인수}}{double COLUMN_DESCRIPTION}
\item{\code{투표수}}{double COLUMN_DESCRIPTION}
\item{\code{무효투표수}}{double COLUMN_DESCRIPTION}
\item{\code{기권수}}{double COLUMN_DESCRIPTION}
\item{\code{후보}}{character COLUMN_DESCRIPTION}
\item{\code{득표수}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://info.nec.go.kr/main/main_load.xhtml}
}
\usage{
education_2014
}
\description{
시도 교육감
}
\details{
제6회 시도 교육감
}
\keyword{datasets}
