% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_president.R
\docType{data}
\name{election_20121219}
\alias{election_20121219}
\title{제18대 대통령선거 2012년}
\format{
투표율( 18455 rows and 8 variables)과 득표율(18455 rows and 20 variable) 데이터프레임 정보를 갖는 리스트(list):
\describe{
\item{\code{투표율}}{-------------------------}
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군명}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{투표구명}}{character COLUMN_DESCRIPTION}
\item{\code{선거인수}}{double COLUMN_DESCRIPTION}
\item{\code{투표수}}{double COLUMN_DESCRIPTION}
\item{\code{무효투표수}}{double COLUMN_DESCRIPTION}
\item{\code{기권수}}{double COLUMN_DESCRIPTION}
\item{\code{득표율}}{-------------------------}
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군명}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{투표구명}}{character COLUMN_DESCRIPTION}
\item{\code{선거인수}}{double COLUMN_DESCRIPTION}
\item{\code{투표수}}{double COLUMN_DESCRIPTION}
\item{\code{박근혜}}{double COLUMN_DESCRIPTION}
\item{\code{문재인}}{double COLUMN_DESCRIPTION}
\item{\code{박종선}}{double COLUMN_DESCRIPTION}
\item{\code{김소연}}{double COLUMN_DESCRIPTION}
\item{\code{강지원}}{double COLUMN_DESCRIPTION}
\item{\code{김순자}}{double COLUMN_DESCRIPTION}
\item{\code{계}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{https://bit.ly/2QRqyGQ}
}
\usage{
election_20121219
}
\description{
2012년 제18대 대통령 선거 결과 (선거관리위원회)
}
\keyword{datasets}
