% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_20170509}
\alias{election_20170509}
\title{제19대 대통령선거 2017년}
\format{
투표율( 18455 rows and 8 variables)과 득표율(18455 rows and 20 variable) 데이터프레임 정보를 갖는 리스트(list):
\describe{
  \item{\code{투표율}}{-------------------------}
  \item{\code{시도명}}{character COLUMN_DESCRIPTION}
  \item{\code{구시군명}}{character COLUMN_DESCRIPTION}
  \item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
  \item{\code{투표구명}}{character COLUMN_DESCRIPTION}
  \item{\code{선거인수}}{double COLUMN_DESCRIPTION}
  \item{\code{투표수}}{double COLUMN_DESCRIPTION}
  \item{\code{무효투표수}}{double COLUMN_DESCRIPTION}
  \item{\code{기권수}}{double COLUMN_DESCRIPTION}
  \item{\code{득표율}}{-------------------------}
  \item{\code{시도명}}{character COLUMN_DESCRIPTION}
  \item{\code{구시군명}}{character COLUMN_DESCRIPTION}
  \item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
  \item{\code{투표구명}}{character COLUMN_DESCRIPTION}
  \item{\code{선거인수}}{double COLUMN_DESCRIPTION}
  \item{\code{투표수}}{double COLUMN_DESCRIPTION}
  \item{\code{문재인}}{double COLUMN_DESCRIPTION}
  \item{\code{홍준표}}{double COLUMN_DESCRIPTION}
  \item{\code{안철수}}{double COLUMN_DESCRIPTION}
  \item{\code{유승민}}{double COLUMN_DESCRIPTION}
  \item{\code{심상정}}{double COLUMN_DESCRIPTION}
  \item{\code{조원진}}{double COLUMN_DESCRIPTION}
  \item{\code{오영국}}{double COLUMN_DESCRIPTION}
  \item{\code{장성민}}{double COLUMN_DESCRIPTION}
  \item{\code{이재오}}{double COLUMN_DESCRIPTION}
  \item{\code{김선동}}{double COLUMN_DESCRIPTION}
  \item{\code{이경희}}{double COLUMN_DESCRIPTION}
  \item{\code{윤홍식}}{double COLUMN_DESCRIPTION}
  \item{\code{김민찬}}{double COLUMN_DESCRIPTION}
  \item{\code{계}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{https://bit.ly/2QRqyGQ}
}
\usage{
election_20170509
}
\description{
2017년 제19대 대통령 선거 결과 (선거관리위원회)
}
\keyword{datasets}
