% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_president.R
\docType{data}
\name{election_20220309}
\alias{election_20220309}
\title{제20대 대통령선거 2022년}
\format{
투표율( 18975 rows and 8 variables)과 득표율(18975 rows and 19 variables) 데이터프레임 정보를 갖는 리스트(list):
\describe{
\item{\code{투표율}}{-------------------------}
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군명}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{투표구명}}{character COLUMN_DESCRIPTION}
\item{\code{선거인수}}{double COLUMN_DESCRIPTION}
\item{\code{투표수}}{double COLUMN_DESCRIPTION}
\item{\code{무효투표수}}{double COLUMN_DESCRIPTION}
\item{\code{기권수}}{double COLUMN_DESCRIPTION}
\item{\code{득표율}}{-------------------------}
\item{\code{더불어민주당이재명}}{double COLUMN_DESCRIPTION}
\item{\code{국민의힘윤석열}}{double COLUMN_DESCRIPTION}
\item{\code{정의당심상정}}{double COLUMN_DESCRIPTION}
\item{\code{기본소득당오준호}}{double COLUMN_DESCRIPTION}
\item{\code{국가혁명당허경영}}{double COLUMN_DESCRIPTION}
\item{\code{노동당이백윤}}{double COLUMN_DESCRIPTION}
\item{\code{새누리당옥은호}}{double COLUMN_DESCRIPTION}
\item{\code{신자유민주연합김경재}}{double COLUMN_DESCRIPTION}
\item{\code{우리공화당조원진}}{double COLUMN_DESCRIPTION}
\item{\code{진보당김재연}}{double COLUMN_DESCRIPTION}
\item{\code{통일한국당이경희}}{double COLUMN_DESCRIPTION}
\item{\code{한류연합당김민찬}}{double COLUMN_DESCRIPTION}
\item{\code{계}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{http://info.nec.go.kr/electioninfo/electionInfo_report.xhtml}
}
\usage{
election_20220309
}
\description{
2022년 제20대 대통령 선거 결과 (선거관리위원회)
}
\keyword{datasets}
