% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_code}
\alias{election_code}
\title{중앙선거관리위원회 선거코드}
\format{
A data frame with 79 rows and 3 variables:
\describe{
\item{선거코드}{문자형, 선거실시일을 코드기준값으로 설정}
\item{선거명}{문자형, 선거명칭}
\item{선거구분}{문자형, 2자리 숫자로 구성된 선거구분코드}
}
}
\source{
\url{https://www.data.go.kr/}
}
\usage{
election_code
}
\description{
공공데이터포털 - 중앙선거관리위원회 선거코드
공공데이터포털 - 중앙선거관리위원회 선거코드
검색어: 중앙선거관리위원회 코드정보
}
\examples{
election_code
}
\keyword{datasets}
