% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_general_election.R
\docType{data}
\name{general_2012}
\alias{general_2012}
\title{제19대 국회의원선거}
\format{
A data frame with 313 rows and 3 variables:
\describe{
\item{\code{시도}}{character COLUMN_DESCRIPTION}
\item{\code{선거구}}{character COLUMN_DESCRIPTION}
\item{\code{data}}{list COLUMN_DESCRIPTION}
}
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=14314}
}
\usage{
general_2012
}
\description{
20120411 국회의원선거
}
\keyword{datasets}
