% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_general_election.R
\docType{data}
\name{general_2020}
\alias{general_2020}
\title{제21대 국회의원선거}
\format{
A data frame with 329 rows and 3 variables:
\describe{
\item{\code{시도}}{character 시도명 2자리}
\item{\code{선거구}}{character 시군구 선거구}
\item{\code{data}}{list 유권자, 투표, 득표}
}
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=15052}
}
\usage{
general_2020
}
\description{
20200415 국회의원선거
}
\keyword{datasets}
