% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_local_election.R
\docType{data}
\name{local_edu_gg_20140604}
\alias{local_edu_gg_20140604}
\title{제6회 지방선거 - 교육감 - 경기}
\format{
A data frame with 1826 rows and 14 variables:
\describe{
\item{\code{위원회명}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{구분}}{character COLUMN_DESCRIPTION}
\item{\code{선거인수}}{double COLUMN_DESCRIPTION}
\item{\code{투표수}}{double COLUMN_DESCRIPTION}
\item{\code{조전혁}}{double COLUMN_DESCRIPTION}
\item{\code{정종희}}{double COLUMN_DESCRIPTION}
\item{\code{최준영}}{double COLUMN_DESCRIPTION}
\item{\code{박용우}}{double COLUMN_DESCRIPTION}
\item{\code{김광래}}{double COLUMN_DESCRIPTION}
\item{\code{이재정}}{double COLUMN_DESCRIPTION}
\item{\code{계}}{double COLUMN_DESCRIPTION}
\item{\code{무효_투표수}}{double COLUMN_DESCRIPTION}
\item{\code{기권수}}{double COLUMN_DESCRIPTION}
}
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=14979}
}
\usage{
local_edu_gg_20140604
}
\description{
경기도 읍면동별 교육감
}
\keyword{datasets}
