% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_local_election.R
\docType{data}
\name{local_edu_seoul_20140604}
\alias{local_edu_seoul_20140604}
\title{제6회 지방선거 - 교육감 - 서울}
\format{
A data frame with 25 rows and 2 variables:
\describe{
\item{\code{시군구명}}{character COLUMN_DESCRIPTION}
\item{\code{data}}{list COLUMN_DESCRIPTION}
}
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=14979}
}
\usage{
local_edu_seoul_20140604
}
\description{
서울시 구시군별 교육감
}
\keyword{datasets}
