% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_local_election.R
\docType{data}
\name{local_sgg_20140604}
\alias{local_sgg_20140604}
\title{제6회 지방선거 - 구시군의장}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{data}}{list 경기도, 서울특별시}
}
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=14979}
}
\usage{
local_sgg_20140604
}
\description{
20140604 지방선거 구시군의장
}
\keyword{datasets}
