% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_local_election.R
\docType{data}
\name{local_sgg_20180613}
\alias{local_sgg_20180613}
\title{제7회 지선 구시군의장}
\format{
A data frame with 226 rows and 3 variables:
\describe{
\item{\code{시도}}{character COLUMN_DESCRIPTION}
\item{\code{선거구명}}{character COLUMN_DESCRIPTION}
\item{\code{data}}{list COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
local_sgg_20180613
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
