% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_local_election.R
\docType{data}
\name{local_sido_20180613}
\alias{local_sido_20180613}
\title{제7회 지방선거 - 시도지사}
\format{
A data frame with 17 rows and 2 variables:
\describe{
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{data}}{list COLUMN_DESCRIPTION}
}
}
\source{
\url{https://www.nec.go.kr/site/nec/ex/bbs/View.do?cbIdx=1129&bcIdx=14979}
}
\usage{
local_sido_20180613
}
\description{
20180613 지방선거 시도지사
}
\keyword{datasets}
