% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{make_dataframe_clean}
\alias{make_dataframe_clean}
\title{make_dataframe_clean}
\usage{
make_dataframe_clean(raw_df)
}
\arguments{
\item{raw_data}{dataframe or tibble}
}
\value{
clean dataframe
}
\description{
국회의원선거 wide 형태를 unnest 가능한 long 형태 변환
}
