% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOC_NEC_master_code.R
\docType{data}
\name{nec_sgg_code}
\alias{nec_sgg_code}
\title{시도, 구시군 마스터 코드}
\format{
A data frame with 17 rows and 3 variables:
\describe{
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{시도코드}}{character COLUMN_DESCRIPTION}
\item{\code{data}}{list COLUMN_DESCRIPTION}
}
}
\source{
\url{http://info.nec.go.kr/electioninfo/electionInfo_report.xhtml}
}
\usage{
nec_sgg_code
}
\description{
시도, 구시군 4자리 마스터 코드
}
\keyword{datasets}
