% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_go_kr_voters.R
\docType{data}
\name{voter_emd}
\alias{voter_emd}
\title{선거별 구시군, 읍면동별 유권자수}
\format{
A data frame with 49653 rows and 27 variables:
\describe{
\item{\code{선거코드}}{character COLUMN_DESCRIPTION}
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군명}}{character COLUMN_DESCRIPTION}
\item{\code{결과순서}}{character COLUMN_DESCRIPTION}
\item{\code{선거ID}}{character COLUMN_DESCRIPTION}
\item{\code{시도명}}{character COLUMN_DESCRIPTION}
\item{\code{구시군명}}{character COLUMN_DESCRIPTION}
\item{\code{읍면동명}}{character COLUMN_DESCRIPTION}
\item{\code{투표구수}}{character COLUMN_DESCRIPTION}
\item{\code{인구수}}{character COLUMN_DESCRIPTION}
\item{\code{인구수(재외국민)}}{character COLUMN_DESCRIPTION}
\item{\code{인구수(외국인)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(계)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(계_재외국민)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(계_외국인)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(남)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(남_재외국민)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(남_외국인)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(여)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(여_재외국민)}}{character COLUMN_DESCRIPTION}
\item{\code{확정선거인수(여_외국인)}}{character COLUMN_DESCRIPTION}
\item{\code{거소투표 신고인명부 등재자수(계)}}{character COLUMN_DESCRIPTION}
\item{\code{거소투표 신고인명부 등재자수(계_재외국민)}}{character COLUMN_DESCRIPTION}
  \item{\code{거소투표 신고인명부 등재자수(남)}}{character COLUMN_DESCRIPTION}
\item{\code{거소투표 신고인명부 등재자수(남_재외국민)}}{character COLUMN_DESCRIPTION}
  \item{\code{거소투표 신고인명부 등재자수(여)}}{character COLUMN_DESCRIPTION}
\item{\code{거소투표 신고인명부 등재자수(여_재외국민)}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
voter_emd
}
\description{
선거별 구시군, 읍면동별 유권자수
}
\keyword{datasets}
